+++ src/cache.d.ts
@@ -1,0 +1,1 @@
+export default function useDebounce<T>(value: T, delay: number): T;
