+++ src/callback.d.ts
@@ -1,0 +1,5 @@
+export default function useDebouncedCallback<T extends (...args: any[]) => any>(
+  callback: T,
+  deps: ReadonlyArray<any>,
+  delay: Number
+): T;
