--- src/index.d.ts
@@ -1,1 +1,0 @@
-export function useDebounce<T>(value: T, delay: number): T;
